+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0-alpha.2](https://bitbucket.org/coveord/ui-kit/compare/v0.5.0-alpha.1...v0.5.0-alpha.2) (2021-02-02)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.5.0-alpha.1](https://bitbucket.org/coveord/ui-kit/compare/v0.5.0-alpha.0...v0.5.0-alpha.1) (2021-02-01)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.5.0-alpha.0](https://bitbucket.org/coveord/ui-kit/compare/v0.4.0...v0.5.0-alpha.0) (2021-02-01)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
 # [0.4.0](https://bitbucket.org/coveord/ui-kit/compare/v0.4.0-alpha.11...v0.4.0) (2021-01-29)
 
 **Note:** Version bump only for package @coveo/quantic