+++ packages/samples/headless-react/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0-alpha.5](https://github.com/coveo/ui-kit/compare/v0.5.0-alpha.4...v0.5.0-alpha.5) (2021-02-02)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+# [0.5.0-alpha.4](https://github.com/coveo/ui-kit/compare/v0.5.0-alpha.3...v0.5.0-alpha.4) (2021-02-02)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+# [0.5.0-alpha.3](https://github.com/coveo/ui-kit/compare/v0.5.0-alpha.2...v0.5.0-alpha.3) (2021-02-02)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+# [0.5.0-alpha.2](https://github.com/coveo/ui-kit/compare/v0.5.0-alpha.1...v0.5.0-alpha.2) (2021-02-02)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+# [0.5.0-alpha.1](https://github.com/coveo/ui-kit/compare/v0.5.0-alpha.0...v0.5.0-alpha.1) (2021-02-01)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+# [0.5.0-alpha.0](https://github.com/coveo/ui-kit/compare/v0.4.0...v0.5.0-alpha.0) (2021-02-01)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
 # [0.4.0](https://github.com/coveo/ui-kit/compare/v0.4.0-alpha.11...v0.4.0) (2021-01-29)
 
 **Note:** Version bump only for package headless-react